+++ src/main/java/charity/pejvak/coinbox/repository/CoinBoxRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CoinBoxRepository extends JpaRepository<CoinBox,Integer> {
+public interface CoinBoxRepository extends JpaRepository<CoinBox,Long> {
 }
